// src/http/models/listCredentialClient200Item.ts
/**
 * Generated by orval v7.6.0 🍺
 * Do not edit manually.
 * onec
 * OpenAPI spec version: 1.0.0
 */
import type { ListCredentialClient200ItemCredentials } from './listCredentialClient200ItemCredentials';
import type { ListCredentialClient200ItemClients } from './listCredentialClient200ItemClients';

export type ListCredentialClient200Item = {
  id: string;
  credentials: ListCredentialClient200ItemCredentials;
  clients: ListCredentialClient200ItemClients;
};